Add Tests for database commands

+++ src/main/java/ua/com/juja/sqlcmd/controller/command/DropDatabase.java
@@ -3,6 +3,8 @@ package ua.com.juja.sqlcmd.controller.command;
 import ua.com.juja.sqlcmd.model.DatabaseManager;
 import ua.com.juja.sqlcmd.view.View;
 
+import java.util.Set;
+
 public class DropDatabase implements Command {
     private final View view;
     private final DatabaseManager manager;
@@ -19,11 +21,15 @@ public class DropDatabase implements Command {
 
     @Override
     public void process(String command) {
-        if (validate(command)) {
+        if (validate(command) && confirm(command)) {
             String[] data = command.split("\\|");
             String databaseName = data[1];
-            manager.dropDatabase(databaseName);
-            view.write(String.format("Database '%s' dropped successfully", databaseName));
+            if (databaseExist(databaseName)) {
+                manager.dropDatabase(databaseName);
+                view.write(String.format("Database '%s' dropped successfully", databaseName));
+            } else {
+                view.write(String.format("Database '%s' not exist. Operation cancelled.", databaseName));
+            }
         }
     }
 
@@ -45,5 +51,25 @@ public class DropDatabase implements Command {
                             "your command: %s", format(), command));
         }
         return true;
+    }
+
+    private boolean confirm(String command) {
+        String[] data = command.split("\\|");
+        String databaseName = data[1];
+        view.write(String.format("To confirm drop database '%s' type 'yes'.", databaseName));
+        if ("yes".equals(view.read().trim())) {
+            return true;
+        } else {
+            view.write(String.format("Drop database '%s' cancelled.", databaseName));
+            return false;
+        }
+    }
+
+    private boolean databaseExist(String databaseName) {
+        Set<String> databases = manager.getDatabasesNames();
+        if (databases.contains(databaseName)) {
+            return true;
+        }
+        return false;
     }
 }
